requirements.txt
fastapi~=0.135.2
starlette~=1.0.0
pydantic~=2.12.5
pyjwt
passlib